Move raised file limit patch to PatchIO
